feat(impersonate): add `random`
