+++ Cargo.toml
@@ -39,6 +39,7 @@ webpki-root-certs = "0.26.8"
 http-body-util = "0.1.3"
 http = "1.3.1"
 mime = "0.3.17"
+rand = "0.9.0"
 
 [profile.release]
 codegen-units = 1
